Wrap Firebase in Dynamic Framework
This allows us to have it in multiple frameworks without conflicting
copies. It includes whichever Firebase Cocoapods you install into it
and firebase auth. I would rather not have auth in it but since it
depends on the Firebase pod, it causes conflicts when it's only in the
main app.